requirements.txt
luma.oled===3.13.0
luma.emulator===1.5.0
pynput===1.7.6